Changed exit code type

+++ include/microv5/archcis/files.h
@@ -5,7 +5,7 @@ std::string fReadStringFromFile(std::string fileName) {
 
 }
 
-short fWriteStringToFile(std::string fileName, std::string fileContent ) {
+int fWriteStringToFile(std::string fileName, std::string fileContent ) {
 
 }
 
@@ -13,6 +13,6 @@ bool fFileExists(std::string fileName) {
   
 }
 
-short fRenameFile(std::string original, std::string renameTo) {
+int fRenameFile(std::string original, std::string renameTo) {
 
 }
